NovaGUI: NovaConfig: configure node profiles in GUI
Node profiles now selectable through drop down boxes, whenever a profile
is renamed, deleted, or created loadAllNodes() needs to be called to
populate the profile drop down boxes. 
Fixed a bug in doppelganger node naming, it was using the name
"Doppelganger" but was using the assigned IP as the key. It now uses it's
name as the key like all other node types, which is "Doppelganger"

New function 'UpdateNodeTypes()' checks every node's name and key
against the mode of their current profile and  either verifies or renames the
node.  Use this function anytime a node might have a new mode/type due to
profile changes.  This is separate from loadAllNodes which simply displays
the contents of the NodeTable, update actually modifies the table.