UI_WEB: Fixing profile deletiong bug and reverting delete nodes debug stuff
Fixed the bug where when attempting to delete a profile and it's parent, 
the error dialog will show up. Now it just calls the delete, making a check
inside the m_profiles hashMap before to make sure the node to be 
deleted is there.

Also, reverted the stuff I did for making bulk node deletion better than it 
is now. May just want to call deleteSelectedItems() twice in the onClick 
event for the delete button just to provide the illusion of a working 
button, though the error will still be there. I have a question up on 
StackOverflow right now to address the issue, but if that doesn't end up
helping we're going to have to either modify Dojo source (yuck) or just
not use a datagrid.

Issue: #354